Move parseIncomingMessage to tests/helpers folder
