+++ packages/core/characterisation-tests/helpers/processMessageCore.ts
@@ -12,8 +12,8 @@ import { PncApiError } from "../../lib/pnc/PncGateway"
 import CorePhase1 from "../../phase1/phase1"
 import CorePhase2 from "../../phase2/phase2"
 import CorePhase3 from "../../phase3/phase3"
-import parseIncomingMessage from "../../tests/helpers/comparison/parseIncomingMessage"
 import MockPncGateway from "../../tests/helpers/MockPncGateway"
+import parseIncomingMessage from "../../tests/helpers/parseIncomingMessage"
 import generateMockPncQueryResult from "./generateMockPncQueryResult"
 
 const createMockPncGateway = (
