+++ packages/core/phase1/phase1.comparison.test.ts
@@ -19,9 +19,9 @@ import { clearDatabase, disconnectDb, sql } from "../tests/helpers/comparison/Co
 import generateMockPncQueryResultFromAho from "../tests/helpers/comparison/generateMockPncQueryResultFromAho"
 import getComparisonTests from "../tests/helpers/comparison/getComparisonTests"
 import getPncQueryTimeFromAho from "../tests/helpers/comparison/getPncQueryTimeFromAho"
-import parseIncomingMessage from "../tests/helpers/comparison/parseIncomingMessage"
 import { matchingExceptions } from "../tests/helpers/comparison/summariseMatching"
 import MockPncGateway from "../tests/helpers/MockPncGateway"
+import parseIncomingMessage from "../tests/helpers/parseIncomingMessage"
 import sortExceptions from "../tests/helpers/sortExceptions"
 import sortTriggers from "../tests/helpers/sortTriggers"
 import phase1 from "./phase1"
