+++ packages/core/phase3/phase3.comparison.test.ts
@@ -5,7 +5,7 @@ import "../tests/helpers/setEnvironmentVariables"
 import { AuditLogEventSource } from "@moj-bichard7/common/types/AuditLogEvent"
 import "jest-xml-matcher"
 
-import type { ParseIncomingMessageResult } from "../tests/helpers/comparison/parseIncomingMessage"
+import type { ParseIncomingMessageResult } from "../tests/helpers/parseIncomingMessage"
 import type { Phase3E2eComparison } from "../tests/types/ComparisonFile"
 import type { PncException } from "../types/Exception"
 import type Phase3Result from "./types/Phase3Result"
@@ -19,8 +19,8 @@ import { clearDatabase, disconnectDb, insertRecords, sql } from "../tests/helper
 import getComparisonTests from "../tests/helpers/comparison/getComparisonTests"
 import normalisePncOperations from "../tests/helpers/comparison/normalisePncOperations"
 import normaliseXml from "../tests/helpers/comparison/normaliseXml"
-import parseIncomingMessage from "../tests/helpers/comparison/parseIncomingMessage"
 import MockPncGateway from "../tests/helpers/MockPncGateway"
+import parseIncomingMessage from "../tests/helpers/parseIncomingMessage"
 import sortTriggers from "../tests/helpers/sortTriggers"
 import { isPncUpdateDataset } from "../types/PncUpdateDataset"
 import { isPncLockError } from "./exceptions/generatePncUpdateExceptionFromMessage"
