+++ packages/core/tests/helpers/parseIncomingMessage.ts
@@ -1,11 +1,11 @@
-import type { AnnotatedHearingOutcome } from "../../../types/AnnotatedHearingOutcome"
-import type { PncUpdateDataset } from "../../../types/PncUpdateDataset"
+import type { AnnotatedHearingOutcome } from "../../types/AnnotatedHearingOutcome"
+import type { PncUpdateDataset } from "../../types/PncUpdateDataset"
 
-import { parseAhoXml } from "../../../lib/parse/parseAhoXml"
-import { parsePncUpdateDataSetXml } from "../../../lib/parse/parsePncUpdateDataSetXml"
-import parseSpiResult from "../../../lib/parse/parseSpiResult"
-import transformSpiToAho from "../../../lib/parse/transformSpiToAho"
-import getMessageType from "../../../phase1/lib/getMessageType"
+import { parseAhoXml } from "../../lib/parse/parseAhoXml"
+import { parsePncUpdateDataSetXml } from "../../lib/parse/parsePncUpdateDataSetXml"
+import parseSpiResult from "../../lib/parse/parseSpiResult"
+import transformSpiToAho from "../../lib/parse/transformSpiToAho"
+import getMessageType from "../../phase1/lib/getMessageType"
 
 export type ParseIncomingMessageResult = HearingOutcomeResult | PncUpdateDatasetResult | SPIResultsResult
